requirements.txt
PySide6
requests
fontTools
UnityPy
PyCriCodecsEx
playsound3
Levenshtein
pyphen
selenium
webdriver-manager
Pillow
apsw